+++ tests/Persisters/FilesystemTest.php
@@ -2,22 +2,45 @@
 
 namespace Rubix\Tests\Persisters;
 
+use Rubix\ML\Persistable;
 use Rubix\ML\Persisters\Persister;
 use Rubix\ML\Persisters\Filesystem;
+use Rubix\ML\Classifiers\DummyClassifier;
 use PHPUnit\Framework\TestCase;
 
 class FilesystemTest extends TestCase
 {
+    protected $persistable;
+
     protected $persister;
 
     public function setUp()
     {
-        $this->persister = new Filesystem('example.temp', true);
+        $this->persistable = new DummyClassifier();
+
+        $this->persister = new Filesystem(__DIR__ . '/test.model', true);
     }
 
     public function test_build_persister()
     {
         $this->assertInstanceOf(Filesystem::class, $this->persister);
         $this->assertInstanceOf(Persister::class, $this->persister);
+    }
+
+    public function test_save_restore_and_delete()
+    {
+        $this->assertFalse(file_exists(__DIR__ . '/test.model'));
+
+        $this->persister->save($this->persistable);
+
+        $this->assertFileExists(__DIR__ . '/test.model');
+
+        $model = $this->persister->restore();
+
+        $this->assertInstanceOf(Persistable::class, $model);
+
+        $this->persister->delete();
+
+        $this->assertFalse(file_exists(__DIR__ . '/test.model'));
     }
 }